[S09P31B108-360] remove : 대본 전체 리스트 조회 쿼리 생성

+++ backend/src/main/java/com/doran/record_book/dto/res/RecordBookResDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RecordBookResDto {
-    private List<Integer> totalScriptList;
+    private List<Long> totalScriptList;
 
     private List<BookDto> bookList;
